Add zuul jobs for tox and linters
This commit will enable the checking of ansible and other
linters in this repo. This is more ground work for adding
additional zuul jobs that will be used by other repos.

Add bashate for checking shell scripts, this is added since
we will be adding shell scripts to the zuul-jobs repo as a
common location for CI related checks

Adding the setup* for tox-docs checking.

Change-Id: Ie14b95d9931fdef430d9088fc7a5e80e7266e93a
Signed-off-by: Saul Wold <sgw@linux.intel.com>