Add Ansible Playbook to build specfiles via Zuul
This initial playbook and roles builds both CentOS and openSUSE
specfiles. The build occurs on OBS and is currently done in a
sandbox area in my home project [0]. The job currently just
starts the OBS job, it does not currently check the build
results, this will come later.

This has been tested with fault so far. The zuul template will
be added to each repo as specfiles are updated.

Added support to automagically build _service files.

[0] https://build.opensuse.org/project/subprojects/home:saulwold

zuul.d/projects needs to use the stx-zuul-jobs-linters to set
ANSIBLE_ROLES_PATH correctly

Change-Id: I6490030894b5887f33a55b1e66e374bcd50960d5
Signed-off-by: Saul Wold <sgw@linux.intel.com>