Rename common jobs & yaml files
Adopt a new naming convention for the constructs defined in this repo:
they are all to be prefixed with "starlingx-common-".

Keep old jobs and project templates, but rename the relevant files to
"legacy-*.yaml". This is necessary to avoid breakage in downstream
projects and branches.

Define 2 jobs with the new naming convention,
starlingx-common-tox-{linters,pep8}.

Story: 2011326
Task: 51561

Change-Id: I4e9339fb9fe952db2c3a9d105282fb7847521f8b
Signed-off-by: Davlet Panech <davlet.panech@windriver.com>